refactoring and putting my code in
